[SDK-75] The PropertiesFileSettingsProvider can handle parentfiles properly
The PropertiesFileSettingsProvider can now handle parent property files. It doesn't safe the properties of the parents in the defaults anymore, it is flat-loaded. That means that every property of the child property file and of all ancestors are safed in one property, that is stored in the Settings class. Each property value for a given key x of a parent will be overriden by its childs property value for the given key x. Furthermore the saveSettings method has been deleted in each SettingsProvider and the SettingsProvider class itself.