[SDK-76] Second possible memory leak fixed and performance increase
In the JobImpl the method mustBeSuppresed() removed activites aswell, same problem as before. It has been fixed by outsourcing the problematic code in another private method called removeNotRunningActivities(). It is used by the addToLogMessageAndCleanup(..) method aswell. The performance has been increased by avoiding to iterate the activities a second time. Instead it is checked if the activity has a parent and if so, delete it by the method removeNotRunningChildActivity(..).