[SDK-91] DataMasking is now added to the JobImpl and the ActivityImpl
Instead of masking each tracedata, extract and event that is created, now some of the fields and attributes of each Activity will be masked during creation/ after using a setter for each of the discussed values. For the JobImpl it is a bit different, it is not masked on initialisation, but on a flush() call. A test is provided that tests if the values and only the discussed values are masked. (If more values are added later to a new MessageFormat, those aren't tested by the test!) Furthermore a setter was added for each businessStart and businessEnd, because they were sent, but couldn't be filled.