[SDK-97] The mapping has changed a bit and invalid input is handled correctly
ActivityImpl.setEventStatus(..) now doesn't just take the eventStatus as input and puts it in the Activity as EventStatus and maps it plain as the JobStatus aswell, but it tries to map it to a valid EventStatus first (null is a valid EventStatus) and sets the status after that to the activity's event. The Job gets a jobStatus update if and only if the eventStatus = SUCCESS, WARNING or ERROR. Otherwise it stays the same. If the input of setEventStatus is invalid, an NjamsSdkRuntimeException will be thrown now.