[SDK-84] Added properties for minimum thread count and max idle time for sender threads.
There are now two new properties that can be set within the settings: njams.client.sdk.minqueuelength and njams.client.sdk.senderthreadidletime. The first one is for defining the core threads that won't be removed even if they are idle. The second property defines the time the not-core-threads can be idle before they are removed. Defaults are 1 core thread and 10000 ms idle time. If more threads are needed in runtime, they will be created and will be removed if they are idle (after the idle time has been exceeded).