[SDK-132] Changed the TraceMessage format
If the TraceMessage is sent with after each TracePoint that is changed, it would flood the connection. Therefore, the CleanTracepointsTask creates and fills a TraceMessageBuilder while it searches for expired TracePoints, and when it has finished, it builds and sends the TraceMessage where all Processes and their Activities are safed, whose Tracepoints are expired.