[SDK-150] Created NjamsService, NjamsServiceLoader, NjamsServiceDummyCache and ServiceFactory and concrete implementations of it
The NjamsService is an interface with getName() to identify the name of the service. It is implemented by the ConfigurationProvider, SettingsProvider, Sender and Receiver. You can create for example an SettingsProviderFactory and call getInstance() to get an instance of an SettingsProvider. This is done by the ServiceFactory that is extended by the SettingsProviderFactory. It creates a new NjamsServiceDummyCache where dummies of all possible services of the particular Service are hold, for example (JmsSender, CloudSender, HttpSender... in one NjamsServiceDummyCache). From one of these dummies, a real instance will be created by Class.newInstance(). The cache itself therefore looks in its internal map, if a Service to this name already exists. If so, return it. Otherwise, tell the NjamsServiceLoader to get the particular Service. If the service isn't in the classpath, an exception will be thrown.