[SDK-155] The JMS Connection won't be released because of a ResourceAllocationException
If a connection is already established and a publisher has been created aswell, send(..) can be called. It tries to send a TextMessage to the Jms Provider, but if a ResourceAllocationException is thrown while sending, the thread waits a short period of time and tries to send again until it is either successfully sent or a threshold of tries is exceeded. Then it tries to reconnect again.